% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions-integrate-GloptiPolyR.R
\name{coef_name_to_array_index}
\alias{coef_name_to_array_index}
\title{Get positions for monomial coefficients}
\usage{
coef_name_to_array_index(coefficients_name, k)
}
\arguments{
\item{coefficients_name}{string vector of shape (1, p); it specifies the
coefficient names following the polym pattern, e.g.,   
the name for x1*x2^2 in a 3-variable model is "1.2.0"}

\item{k}{integer scalor; it specifies the number of variables}
}
\value{
integer matrix of shape (p, k); its (i, j) element speficies the
        (power + 1) value of the jth variable in the ith monomial term,
        (power + 1) accommodating the zero power; its ith row specifies
        the position of the coefficient of the ith nomomial term in the 
        multi-dimensional array of the GloptiPolyR solver
}
\description{
Get positions for monomial coefficients
}
