% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cubic_5D}
\alias{cubic_5D}
\title{Simulated dataset based on a cubic function (5 factors)}
\format{A list consisting of 2 components:
\describe{
  \item{design_matrix}{300 design points generated by a HLS design within R}
  \item{responses}{300 noisy responses simulated at the design points with sigma = 2}
}}
\usage{
cubic_5D
}
\description{
The dataset is simulated from the following function 
\deqn{
f(x) =  10 - (x_1-1.5)^2 - (x_2-2)^2 - (x_3-2.5)^2 - (x_4-3)^2- (x_5-3.5)^2 
+ 0.1 x_1^3 - 0.1 x_2^3 - 0.1 x_3^3 - 0.1 x_4^3 - 0.1 x_5^3 + x_2 x_4 - x_3 x_4
}
defined in the region R = \{ 0 <= x_i <= 5, for i = 1, ...,  5\}, with its maximum at 
(2.28, 2.44, 1.02, 2.65, 2.54).
The sample locations are generated via a Hyper Latin Square (HLS) design within R,
and the noisy level is sigma = 2.
}
\keyword{datasets}
