% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions-scaling.R
\name{decode_orthogonal}
\alias{decode_orthogonal}
\title{Decode using the orthogonal convention}
\usage{
decode_orthogonal(X, lb, ub)
}
\arguments{
\item{X}{numeric matrix of shape (N, k); X specifies the points in
original units}

\item{lb}{numeric vector of shape (1, k); lb specifies the lower bounds for
the k variables}

\item{ub}{numeric vector of shape (1, k); ub specifies the upper bounds for 
the k variables}
}
\value{
numeric matrix of shape (N, k); res specifies the points in
        original units
}
\description{
Decode using the orthogonal convention
}
