% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions-plot-CR.R
\name{draw_2D_CR}
\alias{draw_2D_CR}
\title{Draw a 2D plot for the optima confidence region}
\usage{
draw_2D_CR(boot_optima, boost_optimum, xlab, ylab, xlim, ylim, main)
}
\arguments{
\item{boot_optima}{numeric matrix of shape ((1 - alpha)B, 2); it contains the
(1 - alpha)B bootstrap optima projected onto a 2D plane}

\item{boost_optimum}{numeric vector of shape (1, 2); the bootsted optimum
computed by taking the column average of boot_optima}

\item{xlab}{string; it specifies the lable of the horizontal axis}

\item{ylab}{string; it specifies the lable of the vertical axis}

\item{xlim}{numeric vector of shape (1, 2); it specifies the lower and upper 
limits of the horizontal axis}

\item{ylim}{numeric vector of shape (1, 2); it specifies the lower and upper 
limits of the vertical axis}

\item{main}{string; it specifies the title of the output figure}
}
\value{
a figure displaying the confidence region of the true optimum,
        along with the boostrap optima and the boosted optimum,
        projected onto a 2D plane
}
\description{
Draw a 2D plot for the optima confidence region
}
