% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions-plot-CR.R
\name{draw_2D_CRs}
\alias{draw_2D_CRs}
\title{Draw pair-wise projected 2D plots for the optima confidence region}
\usage{
draw_2D_CRs(boot_optima, boost_optimum, lb, ub)
}
\arguments{
\item{boot_optima}{numeric matrix of shape ((1 - alpha)B, 2); it contains the
(1 - alpha)B bootstrap optima projected onto a 2D plane}

\item{boost_optimum}{numeric vector of shape (1, 2); the bootsted optimum
computed by taking the column average of boot_optima}

\item{lb}{numeric vector of shape (1, k); lb specifies the lower bounds for
the k variables}

\item{ub}{numeric vector of shape (1, k); ub specifies the upper bounds for 
the k variables}
}
\value{
a figure displaying the confidence region of the true optimum,
        projected onto each pairwise-variable planes
}
\description{
Draw pair-wise projected 2D plots for the optima confidence region
}
