% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions-scaling.R
\name{encode_orthogonal}
\alias{encode_orthogonal}
\title{Encode using the orthogonal convention}
\usage{
encode_orthogonal(Xi, lb, ub)
}
\arguments{
\item{Xi}{numeric matrix of shape (N, k); Xi specifies the points in
original units}

\item{lb}{numeric vector of shape (1, k); lb specifies the lower bounds for
the k variables}

\item{ub}{numeric vector of shape (1, k); ub specifies the upper bounds for 
the k variables}
}
\value{
numeric matrix of shape (N, k); it specifies the points in the
        (-1, 1) coded units
}
\description{
Encode using the orthogonal convention
}
