% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions-integrate-GloptiPolyR.R
\name{fit_polym}
\alias{fit_polym}
\title{Fit polynomial models}
\usage{
fit_polym(X, y, degree)
}
\arguments{
\item{X}{numeric matrix of shape (N, k); N is the sample size; k is the 
number of variables, which can be 2, 3, 4 and 5; X specifies the 
design matrix}

\item{y}{numeric vector of shape (N, 1); y specifies the responses}

\item{degree}{integer scalor; degree specifies the order of the polynomial 
model, which can be 2 or 3}
}
\value{
an object of class "lm"
}
\description{
Formulas are entered via the polym function.
Polym automatically formulates the full polynomial basis given its degree
and names coefficients with 0/1 indicators, e.g. the name for term 
x1*x2^2 in a 3 variables model is "1.2.0".
This naming pattern can be used to position the coefficients in the 
multi-dimentional arrays for the GloptiPolyR solver, with the help of 
string-manipulation using regular expression
}
