% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions-integrate-GloptiPolyR.R
\name{Gloptipolym}
\alias{Gloptipolym}
\title{Optimize fitted polynomial functions via GloptiPolyR}
\usage{
Gloptipolym(coefficients, k, degree, lb, ub, maximization)
}
\arguments{
\item{coefficients}{numeric vector of shape (1, p); it specifies the the 
coefficients of an "lm" objected formulated with the 
polym function}

\item{k}{integer scalor; it specifies the number of variables}

\item{degree}{integer scalor; degree specifies the order of the polynomial 
model, which can be 2 or 3}

\item{lb}{numeric vector of shape (1, k); lb specifies the lower bounds for
the k variables}

\item{ub}{numeric vector of shape (1, k); ub specifies the upper bounds for 
the k variables}

\item{maximization}{boolean scalor; if specifies whether the algorithm 
computes the confidence region for the maxima or minima}
}
\value{
the optimal solution and its corresponding objective value
}
\description{
Optimize fitted polynomial functions via GloptiPolyR
}
