% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GloptiPolyRegion.R
\name{GloptiPolyRegion}
\alias{GloptiPolyRegion}
\title{Compute confidence regions for polynomial models' optima}
\usage{
GloptiPolyRegion(X, y, degree, lb, ub, B = 200, alpha = 0.05,
  maximization = TRUE, verbose = TRUE)
}
\arguments{
\item{X}{numeric matrix of shape (N, k); N is the sample size; k is the 
number of variables, which can be 2, 3, 4 and 5; X specifies the 
design matrix}

\item{y}{numeric vector of shape (N, 1); y specifies the responses}

\item{degree}{integer scalor; degree specifies the order of the polynomial 
model, which can be 2 or 3}

\item{lb}{numeric vector of shape (1, k); lb specifies the lower bounds for
the k variables}

\item{ub}{numeric vector of shape (1, k); ub specifies the upper bounds for 
the k variables}

\item{B}{integer scalor; B specifies the number of bootstrap operations}

\item{alpha}{numeric scalor between 0 and 1; alpha specifies the nominal 
confidence level, 1 - alpha, of the confidence region}

\item{maximization}{boolean scalor; if specifies whether the algorithm 
computes the confidence region for the maxima or minima}

\item{verbose}{boolean scalor; it specifies whether to display running status}
}
\value{
a figure displaying the confidence region of the true optimum,
        projected onto each pairwise-variable planes and a list consisting
        of 2 components: 
        \itemize{
          \item{boot_optima: }{numeric matrix of shape ((1 - alpha)B, k);
                               it contains the (1 - alpha)B bootstrap optima}
          \item{boost_optimum: }{numeric vector of shape (1, k); the boosted 
                                 optimum; computed by taking the column average
                                 of boot_optima}
        }
}
\description{
Compute confidence regions for polynomial models' optima
}
