% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{quad_3D}
\alias{quad_3D}
\title{A simulated dataset based on a quadratic function}
\format{A list consisting of 2 components:
\describe{
  \item{design_matrix}{200 design points generated by a HLS design within R}
  \item{responses}{200 noisy responses simulated at the design points (sigma = 10)}
}}
\usage{
quad_3D
}
\description{
The dataset is simulated from the following function 
\insertCite{BoxDraper}{OptimaRegion}:
\deqn{
f(x) =  57.31 + 1.5 x_1 - 2.13 x_2 + 1.81 x_3
- 7.13 x_1 x_2 - 3.27 x_1 x_3 - 2.73 x_2 x_3 - 4.69 x_1^2 - 6.27 x_2^2 - 5.21 x_3^2,
}
defined in the region R = \{-2 <= x_i <= 2, for i = 1, 2, 3\}, with its maximum at 
(0.46, -0.46, 0.15).
The sample locations are generated via a Hyper Latin Square (HLS) design within R,
and the noisy level is sigma = 10.
}
\references{
{
  \insertAllCited{}
}
}
\keyword{datasets}
