% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GloptiPolyR.R
\name{GloptiPolyR}
\alias{GloptiPolyR}
\title{Optimize a Quadratic or Cubic Polynomial Function}
\usage{
GloptiPolyR(P)
}
\arguments{
\item{P}{A list of list; Each sub-list has 2 components: 
1. a multi-dimensional array corresponding to a objective or constraint function
2. an attribute of the objective or constraint function}
}
\value{
Returns the optimal solution and its corresponding objective value
}
\description{
Optimize a quadratic or cubic polynomial functionin 2 ~ 5 variables 
with bound constraints
\insertCite{DelCastilloCR}{OptimaRegion}.
}
\section{Author(s)}{

Enrique del Castillo \email{exd13@psu.edu},
Peng Chen \email{pfc5098@psu.edu}, 
Adam Meyers \email{akm5733@psu.edu},
John Hunt \email{J.Hunt@exeter.ac.uk} and
James Rapkin \email{jr297@exeter.ac.uk}.
}

\references{
{
 \insertAllCited{}
}
}
