% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{quad_3D}
\alias{quad_3D}
\title{Simulated dataset based on a qudratic function in 3 variables}
\format{An object of class list of length 2
\describe{
  \item{design matrix}{200 design points generated by a HLS design within R}
  \item{responses}{200 noisy responses at the design points (sigma = 10)}
}}
\usage{
quad_3D
}
\description{
The true function is f(x) =  57.31 + 1.5 x_1 - 2.13 x_2 + 1.81 x_3
- 7.13 x_1 x_2 - 3.27 x_1 x_3 - 2.73 x_2 x_3 - 4.69 x_1^2 - 6.27 x_2^2 - 5.21 x_3^2,
defined in the region R = \{-2 <= x_i <= 2, for i = 1, 2, 3\}, with its maximum at 
(0.46, -0.46, 0.15)(Box and Draper 1987, Chapter 9).
The sample locations are generated via a Hyper Latin Square (HLS) design within R,
and the noisy level is sigma = 10.
}
\keyword{datasets}
