% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{quad_3D}
\alias{quad_3D}
\title{Central composite design experiment dataset (3 factors)}
\format{
A data frame with 16 observations on the following 4 variables
\describe{
  \item{x1}{percentage concentration of the 1st constituent, in coded unit}
  \item{x2}{percentage concentration of the 2nd constituent, in coded unit}
  \item{x3}{temperature, in coded unit}
  \item{y}{elasticity of certain polymer}
}
}
\usage{
quad_3D
}
\description{
\insertCite{BoxDraper;textual}{OptimaRegion}'s three factor experimental dataset
}
\references{
{
  \insertAllCited{}
}
}
\keyword{datasets}
